requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
autopep8==2.0.1
Babel==2.11.0
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
defusedxml==0.7.1
dj-database-url==1.2.0
Django==3.2
django-allauth==0.52.0
django-cors-headers==3.13.0
django-crispy-forms==2.0
django-csp==3.7
django-extensions==3.2.1
django-filter==22.1
django-permissions-policy==4.14.0
django-templated-mail==1.1.1
django-userforeignkey==0.5.0
djangorestframework==3.12.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.2
oauthlib==3.2.2
Pillow==9.4.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
py-moneyed==2.0
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
python-decouple==3.7
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14
whitenoise==6.3.0
zipp==3.14.0
